moved db models to db package
