+++ internal/datastore/mongo_store.go
@@ -5,7 +5,7 @@ import (
 	"log"
     "os"
 
-	"github.com/SimplQ/simplQ-golang/internal/models"
+	"github.com/SimplQ/simplQ-golang/internal/models/db"
 
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
@@ -46,7 +46,7 @@ func NewMongoDB() *MongoDB {
     return &new_mongodb
 }
 
-func (mongodb MongoDB) CreateQueue(queue models.Queue) models.QueueId {
+func (mongodb MongoDB) CreateQueue(queue db.Queue) db.QueueId {
     // Set id to empty so its generated by mongoDB
     queue.Id = ""
 
@@ -58,33 +58,33 @@ func (mongodb MongoDB) CreateQueue(queue models.Queue) models.QueueId {
  
     stringId := result.InsertedID.(primitive.ObjectID).Hex()
 
-    return models.QueueId(stringId)
+    return db.QueueId(stringId)
 }
 
-func (mongodb MongoDB) ReadQueue(models.QueueId) models.Queue {
+func (mongodb MongoDB) ReadQueue(db.QueueId) db.Queue {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) PauseQueue(models.QueueId) {
+func (mongodb MongoDB) PauseQueue(db.QueueId) {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) ResumeQueue(models.QueueId) {
+func (mongodb MongoDB) ResumeQueue(db.QueueId) {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) DeleteQueue(models.QueueId) {
+func (mongodb MongoDB) DeleteQueue(db.QueueId) {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) AddTokenToQueue(models.QueueId, models.Token) {
+func (mongodb MongoDB) AddTokenToQueue(db.QueueId, db.Token) {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) ReadToken(models.TokenId) {
+func (mongodb MongoDB) ReadToken(db.TokenId) {
     panic("Not implemented")
 }
 
-func (mongodb MongoDB) RemoveToken(models.TokenId) {
+func (mongodb MongoDB) RemoveToken(db.TokenId) {
     panic("Not implemented")
 }
