--- internal/models/common.go
@@ -1,3 +1,0 @@
-package models
-
-type Id string
