+++ internal/models/db/common.go
@@ -1,0 +1,3 @@
+package db
+
+type Id string
