+++ internal/models/db/queue.go
@@ -1,4 +1,4 @@
-package models
+package db
 
 import (
 	"time"
