+++ internal/models/db/token.go
@@ -1,4 +1,4 @@
-package models
+package db
 
 import (
 	"time"
